+++ client/src/lib/axios.js
@@ -1,6 +1,7 @@
-const axios = require('axios');
+import axios from 'axios';
+import { SERVER_URL } from '../config';
 
-axios.defaults.baseURL = 'http://localhost:3000/api';
+axios.defaults.baseURL = `${SERVER_URL}/api`;
 const request = async ({ method, params = '', data = '' }) => {
   const token = localStorage.getItem('jwt_token');
