+++ client/src/pages/Issues/MarkAs/index.jsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import { Div, ListItem, Name, Modal } from './styled';
-import { updateState } from '../../apis/issue';
-import { IssueContext } from '../../stores/issueStore';
+import { updateState } from '../../../apis/issue';
+import { IssueContext } from '../../../stores/issueStore';
 
 const MarkAs = ({ checked, checkHandler }) => {
   const {
